boilerplate/openshift/golang-osd-operator/py-requirements.txt
pyyaml>=5.3.1
